Link against libm on unixoids by default
Since it is unknown whether compiled source code depends on libm,
libm should be linked against when producing shared objects by default
in order to avoid producing binaries that implicitly depend on libm
being loaded in the importing implementation.  The linker will be
clever enough not to add that dependency if no symbols are needed
from libm.